Upgrade dependencies and rewrite one unit test as a Wiremock test due to HttpClientV2
